refactor: remove unnecessary usnge and spaces
