+++ Streetcode/Streetcode.BLL/DTO/Timeline/TimelineItem/UpdateTimelineItemDTO.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Streetcode.BLL.DTO.Timeline
+namespace Streetcode.BLL.DTO.Timeline.TimelineItem
 {
 
     public class UpdateTimelineItemDTO
