+++ Streetcode/Streetcode.BLL/Mapping/Timeline/HistoricalContextProfile.cs
@@ -1,5 +1,5 @@
 using AutoMapper;
-using Streetcode.BLL.DTO.Timeline;
+using Streetcode.BLL.DTO.Timeline.HistoricalContext;
 using Streetcode.DAL.Entities.Timeline;
 
 namespace Streetcode.BLL.Mapping.Timeline;
