+++ Streetcode/Streetcode.BLL/Mapping/Timeline/TimelineItemProfile.cs
@@ -1,5 +1,6 @@
 using AutoMapper;
 using Streetcode.BLL.DTO.Timeline;
+using Streetcode.BLL.DTO.Timeline.TimelineItem;
 using Streetcode.DAL.Entities.Timeline;
 
 namespace Streetcode.BLL.Mapping.Timeline;
