+++ Streetcode/Streetcode.BLL/MediatR/Timeline/TimelineItem/GetByStreetcodeId/GetTimelineItemsByStreetcodeIdHandler.cs
@@ -2,7 +2,7 @@ using AutoMapper;
 using FluentResults;
 using MediatR;
 using Microsoft.EntityFrameworkCore;
-using Streetcode.BLL.DTO.Timeline;
+using Streetcode.BLL.DTO.Timeline.TimelineItem;
 using Streetcode.BLL.Interfaces.Logging;
 using Streetcode.DAL.Repositories.Interfaces.Base;
 using Streetcode.Resources;
